# Optimize the classpath implementation