--- a/src/compiler/scala/tools/nsc/classpath/ZipArchiveFileLookup.scala
+++ b/src/compiler/scala/tools/nsc/classpath/ZipArchiveFileLookup.scala
@@ -14,18 +14,19 @@ package scala.tools.nsc.classpath
 
 import java.io.{Closeable, File}
 import java.net.URL
+
 import scala.collection.Seq
 import scala.reflect.io.AbstractFile
 import scala.reflect.io.FileZipArchive
 import FileUtils.AbstractFileOps
-import scala.tools.nsc.util.{ClassPath, ClassRepresentation}
+import scala.tools.nsc.util.{ClassRepresentation, EfficientClassPath}
 
 /**
  * A trait allowing to look for classpath entries of given type in zip and jar files.
  * It provides common logic for classes handling class and source files.
  * It's aware of things like e.g. META-INF directory which is correctly skipped.
  */
-trait ZipArchiveFileLookup[FileEntryType <: ClassRepresentation] extends ClassPath with Closeable {
+trait ZipArchiveFileLookup[FileEntryType <: ClassRepresentation] extends EfficientClassPath with Closeable {
   val zipFile: File
   def release: Option[String]
 
@@ -36,54 +37,47 @@ trait ZipArchiveFileLookup[FileEntryType <: ClassRepresentation] extends ClassPa
   private val archive = new FileZipArchive(zipFile, release)
   override def close(): Unit = archive.close()
 
-  override private[nsc] def packages(inPackage: String): Seq[PackageEntry] = {
-    val prefix = PackageNameUtils.packagePrefix(inPackage)
+  override private[nsc] def packages(inPackage: PackageName): Seq[PackageEntry] = {
     for {
       dirEntry <- findDirEntry(inPackage).toSeq
       entry <- dirEntry.iterator if entry.isPackage
-    } yield PackageEntryImpl(prefix + entry.name)
+    } yield PackageEntryImpl(inPackage.entryName(entry.name))
   }
 
-  protected def files(inPackage: String): Seq[FileEntryType] =
+  protected def files(inPackage: PackageName): Seq[FileEntryType] =
     for {
       dirEntry <- findDirEntry(inPackage).toSeq
       entry <- dirEntry.iterator if isRequiredFileType(entry)
     } yield createFileEntry(entry)
 
-  protected def file(inPackage: String, name: String): Option[FileEntryType] =
+  protected def file(inPackage: PackageName, name: String): Option[FileEntryType] =
     for {
       dirEntry <- findDirEntry(inPackage)
       entry <- Option(dirEntry.lookupName(name, directory = false))
       if isRequiredFileType(entry)
     } yield createFileEntry(entry)
 
-  override private[nsc] def hasPackage(pkg: String) = findDirEntry(pkg).isDefined
-  override private[nsc] def list(inPackage: String): ClassPathEntries = {
-    val foundDirEntry = findDirEntry(inPackage)
+  override private[nsc] def hasPackage(pkg: PackageName) = findDirEntry(pkg).isDefined
 
-    foundDirEntry map { dirEntry =>
-      val pkgBuf = collection.mutable.ArrayBuffer.empty[PackageEntry]
-      val fileBuf = collection.mutable.ArrayBuffer.empty[FileEntryType]
-      val prefix = PackageNameUtils.packagePrefix(inPackage)
-
-      for (entry <- dirEntry.iterator) {
-        if (entry.isPackage)
-          pkgBuf += PackageEntryImpl(prefix + entry.name)
-        else if (isRequiredFileType(entry))
-          fileBuf += createFileEntry(entry)
-      }
-      ClassPathEntries(pkgBuf, fileBuf)
-    } getOrElse ClassPathEntries.empty
+  private[nsc] def list(inPackage: PackageName, onPackageEntry: PackageEntry => Unit, onClassesAndSources: ClassRepresentation => Unit): Unit = {
+    findDirEntry(inPackage) match {
+      case Some(dirEntry) =>
+        for (entry <- dirEntry.iterator) {
+          if (entry.isPackage)
+            onPackageEntry(PackageEntryImpl(inPackage.entryName(entry.name)))
+          else if (isRequiredFileType(entry))
+            onClassesAndSources(createFileEntry(entry))
+        }
+      case None =>
+    }
   }
 
-  private def findDirEntry(pkg: String): Option[archive.DirEntry] = {
-    archive.allDirs.get(dottedToPath(pkg))
+  private def findDirEntry(pkg: PackageName): Option[archive.DirEntry] = {
+    Option(archive.allDirs.get(pkg.dirPathTrailingSlash))
   }
 
-  private def dottedToPath(dotted: String): String = {
-    dotted.replace('.', '/') + "/"
-  }
 
   protected def createFileEntry(file: FileZipArchive#Entry): FileEntryType
   protected def isRequiredFileType(file: AbstractFile): Boolean
 }
+