… src/compiler/scala/tools/nsc/util/ClassPath.scala
… src/reflect/scala/reflect/io/ZipArchive.scala
… test/junit/scala/tools/nsc/classpath/AggregateClassPathTest.scala