- Avoid repeated conversion between package dotted names ("com.foo")
    and relative paths ("com/foo/") by threading a `PackageName`
    instance through the aggregate classpath lookup which caches
    the path name.
  - Avoid creating result buffers inside of each element of the
    aggregate classpath by introducing a callback based API that
    lets `AggregateClasspath` use a single buffer.

Before:

```
[info] Benchmark                                                                                  (classpathString)  Mode  Cnt        Score         Error   Units
[info] AggregateClassPathBenchmark.test                                   @/code/scala-2.12.x/sandbox/classpath.txt  avgt   10  2109607.861 ±  339561.922   ns/op
[info] AggregateClassPathBenchmark.test:·gc.alloc.rate                    @/code/scala-2.12.x/sandbox/classpath.txt  avgt   10      732.202 ±      94.633  MB/sec
[info] AggregateClassPathBenchmark.test:·gc.alloc.rate.norm               @/code/scala-2.12.x/sandbox/classpath.txt  avgt   10  2411260.302 ±      50.691    B/op
[info] AggregateClassPathBenchmark.test:·gc.churn.PS_Eden_Space           @/code/scala-2.12.x/sandbox/classpath.txt  avgt   10      792.591 ±    1269.210  MB/sec
[info] AggregateClassPathBenchmark.test:·gc.churn.PS_Eden_Space.norm      @/code/scala-2.12.x/sandbox/classpath.txt  avgt   10  2732481.613 ± 4514800.213    B/op
[info] AggregateClassPathBenchmark.test:·gc.churn.PS_Survivor_Space       @/code/scala-2.12.x/sandbox/classpath.txt  avgt   10        0.019 ±       0.063  MB/sec
[info] AggregateClassPathBenchmark.test:·gc.churn.PS_Survivor_Space.norm  @/code/scala-2.12.x/sandbox/classpath.txt  avgt   10       59.982 ±     203.957    B/op
[info] AggregateClassPathBenchmark.test:·gc.count                         @/code/scala-2.12.x/sandbox/classpath.txt  avgt   10        5.000                counts
[info] AggregateClassPathBenchmark.test:·gc.time                          @/code/scala-2.12.x/sandbox/classpath.txt  avgt   10        5.000                    ms
```

After:

```
[info] Benchmark                                                                                  (classpathString)  Mode  Cnt       Score        Error   Units
[info] AggregateClassPathBenchmark.test                                   @/code/scala-2.12.x/sandbox/classpath.txt  avgt   10  580983.606 ±  14088.052   ns/op
[info] AggregateClassPathBenchmark.test:·gc.alloc.rate                    @/code/scala-2.12.x/sandbox/classpath.txt  avgt   10     405.902 ±      9.647  MB/sec
[info] AggregateClassPathBenchmark.test:·gc.alloc.rate.norm               @/code/scala-2.12.x/sandbox/classpath.txt  avgt   10  371348.000 ±     14.589    B/op
[info] AggregateClassPathBenchmark.test:·gc.churn.PS_Eden_Space           @/code/scala-2.12.x/sandbox/classpath.txt  avgt   10     429.614 ±    351.032  MB/sec
[info] AggregateClassPathBenchmark.test:·gc.churn.PS_Eden_Space.norm      @/code/scala-2.12.x/sandbox/classpath.txt  avgt   10  392532.252 ± 320194.247    B/op
[info] AggregateClassPathBenchmark.test:·gc.churn.PS_Survivor_Space       @/code/scala-2.12.x/sandbox/classpath.txt  avgt   10       0.006 ±      0.030  MB/sec
[info] AggregateClassPathBenchmark.test:·gc.churn.PS_Survivor_Space.norm  @/code/scala-2.12.x/sandbox/classpath.txt  avgt   10       5.595 ±     26.749    B/op
[info] AggregateClassPathBenchmark.test:·gc.count                         @/code/scala-2.12.x/sandbox/classpath.txt  avgt   10       8.000               counts
[info] AggregateClassPathBenchmark.test:·gc.time                          @/code/scala-2.12.x/sandbox/classpath.txt  avgt   10       7.000                   ms
```

Given:

```
$ coursier fetch -p org.apache.spark:spark-repl_2.12:2.4.3 -p org.apache.cassandra:cassandra-all:3.11.4 | tee /code/scala-2.12.x/sandbox/classpath.txt
```