@retronym 2019-08-21
After this change, the bulk of the allocations are in stripping ".class" of file names in `ClassFileEntryImpl.<init>` and constructing nodes within the `indices: HashMap` in `mergeClassesAndSources`.

The majority of CPU is spent doing `HashMap` lookups in `ZipArchiveFileLookup.findDirEntry`.