@hrhino 2019-08-22
Would it be worth it to do `path.subpath(path.length - 1, path.length).toString endsWith ".class"`? In case of deeply-nested filesystem hierarchies.