@som-snytt 2019-09-27
I also hit the `path.endsWith` semantics at the office recently. :ghost: <- I wanted a "boo" emoji for something that pops up like that out of nowhere.